Allow sort parameters not intended for filterable
Some cases may need a more complicated sort ability than filterable
offers. If a "sort" parameter is used, filterable would hijack the
request trying to parse sort_params out of the parameters whether or not
sort_on was being used at all. If the formatting is not identical on the
custom sort, filterable will throw an error similar to:

NoMethodError: undefined method `split' for {...}:ActionController::Parameters

This solution first checks to see if there is a sorted_on field defined
on the class level and does not try to parse sort_params if none have
been made.

Shortfalls: If a custom sort method is being used in conjunction with
filterable's sorting, there may be an issue since sort_on will define
something. Also, however less of a problem, if a custom sort is being
used elsewhere and the same custom sort parameter format is used during
filterables sort, there will be a surprising error.